requirements.txt
Django==3.0.3
asgiref==3.2.3
pip==20.0.2
psycopg2==2.8.4
python-dotenv==0.11.0
pytz==2019.3
setuptools==40.8.0
sqlparse==0.3.0
djangorestframework=3.11.0